rename webpack resolver plugin
