+++ gridsome/lib/webpack/plugins/GridsomeResolverPlugin.js
